Keeps track of the cassette persisters in a hash-like object

Keeps track of the cassette persisters in a hash-like object

persister_fetch(x = "FileSystem", file_name)

Arguments

persisters
A list

name
(character) Persister name, only option right now is "FileSystem"

Details

There's only one option: FileSystem

Private Methods

persister_get()

Gets and sets a named persister

Examples

(aa <- Persisters$new())

#> <Persisters>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     initialize: function (persisters = list(), name = "FileSystem") 
#>     name: FileSystem
#>     persisters: R6ClassGenerator
#>   Private:
#>     persister_get: function () 

aa$name

#> [1] "FileSystem"

aa$persisters

#> <FileSystem> object generator
#>   Public:
#>     file_name: NULL
#>     write_fxn: NULL
#>     content: NULL
#>     path: NULL
#>     write2disk: FALSE
#>     initialize: function (file_name = NULL, write_fxn = NULL, content = NULL, 
#>     get_cassette: function (file_name = NULL) 
#>     is_empty: function () 
#>     set_cassette: function (file_name = NULL, content) 
#>     clone: function (deep = FALSE) 
#>   Private:
#>     storage_location: function () 
#>     absolute_path_to_file: function (x, y) 
#>   Parent env: <environment: namespace:vcr>
#>   Locked objects: TRUE
#>   Locked class: FALSE
#>   Portable: TRUE

yaml_serializer <- aa$persisters$new()
yaml_serializer

#> <FileSystem>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     content: NULL
#>     file_name: NULL
#>     get_cassette: function (file_name = NULL) 
#>     initialize: function (file_name = NULL, write_fxn = NULL, content = NULL, 
#>     is_empty: function () 
#>     path: /var/folders/fc/n7g_vrvn0sx_st0p8lxb3ts40000gn/T//RtmpIoF2H9
#>     set_cassette: function (file_name = NULL, content) 
#>     write_fxn: NULL
#>     write2disk: FALSE
#>   Private:
#>     absolute_path_to_file: function (x, y) 
#>     storage_location: function ()