UnhandledHTTPRequestError

UnhandledHTTPRequestError

Arguments

request
a request

cassette
a cassette, the current cassette

Details

How this error class is used: If record="once" we trigger this.

Users can use vcr in the context of both use_cassette and insert_cassette.

For the former, all requests go through the call_block But for the latter, requests go through webmockr

Where is one place where we can put UnhandledHTTPRequestError that will handle both use_cassette and insert_cassette?

Error situations where this is invoked

record=once AND there's a new request that doesn't match the one in the cassette on disk

in webmockr: if no stub found and there are recorded interactions on the cassette, and record = once, then error with UnhandledHTTPRequestError

but if record != once, then allow it, unless record == none

others?

Examples

vcr_configure(dir = tempdir())

#> <vcr configuration>
#>   Cassette Dir: /var/folders/fc/n7g_vrvn0sx_st0p8lxb3ts40000gn/T//RtmpIoF2H9
#>   Record: once
#>   URI Parser: crul::url_parse
#>   Match Requests on: method, uri
#>   Preserve Bytes?: FALSE
#>   Logging?: FALSE

cassettes()

#> list()

insert_cassette("turtle")

#> CrulAdapter enabled!

#> HttrAdapter enabled!

#> <vcr - Cassette> turtle
#>   Record method: once
#>   Serialize with: yaml
#>   Persist with: FileSystem
#>   Re-record interval (s): 
#>   Clean outdated interactions?: FALSE
#>   update_content_length_header: FALSE
#>   decode_compressed_response: 
#>   allow_playback_repeats: FALSE
#>   allow_unused_http_interactions: 
#>   exclusive: 
#>   preserve_exact_body_bytes: FALSE

request <- Request$new("post", 'https://eu.httpbin.org/post?a=5',
  "", list(foo = "bar"))

err <- UnhandledHTTPRequestError$new(request)
err$request_description()

#> [1] "POST https://eu.httpbin.org/post?a=5"

err$current_matchers()

#> [1] "method" "uri"   

err$match_request_on_headers()

#> [1] FALSE

err$match_request_on_body()

#> [1] FALSE

err$formatted_headers()

#> [1] "    foo: bar"

cat(err$formatted_headers(), "\n")

#>     foo: bar 

cat(err$cassettes_description(), "\n")

#> vcr is currently using the following cassette:
#>   - /var/folders/fc/n7g_vrvn0sx_st0p8lxb3ts40000gn/T//RtmpIoF2H9/turtle.yml
#>     - record_mode: once
#>     - match_requests_on: method, uri 
#> 
#> Under the current configuration vcr can not find a suitable HTTP interaction
#> to replay and is prevented from recording new requests. There are a few ways
#> you can deal with this:
#>  

cat(err$cassettes_list(), "\n")

#> vcr is currently using the following cassette:
#>   - /var/folders/fc/n7g_vrvn0sx_st0p8lxb3ts40000gn/T//RtmpIoF2H9/turtle.yml
#>     - record_mode: once
#>     - match_requests_on: method, uri 

err$formatted_suggestions()

#> [1] "  * If you're surprised vcr is raising this error\n    and want insight about how vcr attempted to handle the request,\n    you can use 'logging' to see more details [1].\n\n  * You can use the :new_episodes record mode to allow vcr to\n    record this new request to the existing cassette [2].\n\n  * If you want vcr to ignore this request (and others like it), you can\n    set an `ignore_request` function [3].\n\n  * The current record mode ('once') does not allow new requests to be recorded\n    to a previously recorded cassette. You can delete the cassette file and re-run\n    your tests to allow the cassette to be recorded with this request [4].\n\n[1] https://ropenscilabs.github.io/http-testing-book/logging\n\n[2] https://ropenscilabs.github.io/http-testing-book/record-modes.html#new_episodes\n\n[3] https://ropenscilabs.github.io/http-testing-book/vcr-configuration#config-ignore-requests\n\n[4] https://ropenscilabs.github.io/http-testing-book/record-modes.html#once\n"

cat(err$format_bullet_point('foo bar', 1), "\n")

#>   * foo bar [2]. 

err$suggestion_for("use_new_episodes")

#> $text
#> [1] "You can use the :new_episodes record mode to allow vcr to"
#> [2] "record this new request to the existing cassette"         
#> 
#> $url
#> [1] "https://ropenscilabs.github.io/http-testing-book/record-modes.html#new_episodes"
#> 

err$suggestions()

#> [[1]]
#> [[1]]$text
#> [1] "If you're surprised vcr is raising this error"                  
#> [2] "and want insight about how vcr attempted to handle the request,"
#> [3] "you can use 'logging' to see more details"                      
#> 
#> [[1]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/logging"
#> 
#> 
#> [[2]]
#> [[2]]$text
#> [1] "You can use the :new_episodes record mode to allow vcr to"
#> [2] "record this new request to the existing cassette"         
#> 
#> [[2]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/record-modes.html#new_episodes"
#> 
#> 
#> [[3]]
#> [[3]]$text
#> [1] "If you want vcr to ignore this request (and others like it), you can"
#> [2] "set an `ignore_request` function"                                    
#> 
#> [[3]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/vcr-configuration#config-ignore-requests"
#> 
#> 
#> [[4]]
#> [[4]]$text
#> [1] "The current record mode ('once') does not allow new requests to be recorded"   
#> [2] "to a previously recorded cassette. You can delete the cassette file and re-run"
#> [3] "your tests to allow the cassette to be recorded with this request"             
#> 
#> [[4]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/record-modes.html#once"
#> 
#> 

err$no_cassette_suggestions()

#> [[1]]
#> [[1]]$text
#> [1] "If you're surprised vcr is raising this error"                  
#> [2] "and want insight about how vcr attempted to handle the request,"
#> [3] "you can use 'logging' to see more details"                      
#> 
#> [[1]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/logging"
#> 
#> 
#> [[2]]
#> [[2]]$text
#> [1] "If you want vcr to record this request and play it back during future test"
#> [2] "runs, you should wrap your test (or this portion of your test) in a"       
#> [3] "`vcr::use_cassette` block"                                                 
#> 
#> [[2]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/intro"
#> 
#> 
#> [[3]]
#> [[3]]$text
#> [1] "If you only want vcr to handle requests made while a cassette is in use,"
#> [2] "configure `allow_http_connections_when_no_cassette = TRUE`. vcr will"    
#> [3] "ignore this request since it is made when there is no cassette"          
#> 
#> [[3]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/vcr-configuration#allow-http-connections-when-no-cassette"
#> 
#> 
#> [[4]]
#> [[4]]$text
#> [1] "If you want vcr to ignore this request (and others like it), you can"
#> [2] "set an `ignore_request` function"                                    
#> 
#> [[4]]$url
#> [1] "https://ropenscilabs.github.io/http-testing-book/vcr-configuration#config-ignore-requests"
#> 
#> 

err$record_mode_suggestion()

#> [1] "delete_cassette_for_once"

err$has_used_interaction_matching()

#> [1] FALSE

err$match_requests_on_suggestion()

#> NULL

# err$construct_message()

# cleanup
unlink(tempdir())