Eject a cassette

Eject a cassette

eject_cassette(cassette = NULL, options = list(),
  skip_no_unused_interactions_assertion = NULL)

Arguments

cassette
(character) a single cassette names to eject

options
(list) a list of options to apply to the eject process

skip_no_unused_interactions_assertion
(logical) If TRUE, this will skip the "no unused HTTP interactions" assertion enabled by the allow_unused_http_interactions = FALSE cassette option. This is intended for use when your test has had an error, but your test framework has already handled it - IGNORED FOR NOW

Value

The ejected cassette if there was one

See also

Examples

vcr_configure(dir = tempdir())

#> <vcr configuration>
#>   Cassette Dir: /var/folders/fc/n7g_vrvn0sx_st0p8lxb3ts40000gn/T//RtmpYxsTis
#>   Record: once
#>   URI Parser: crul::url_parse
#>   Match Requests on: method, uri
#>   Preserve Bytes?: FALSE
#>   Logging?: FALSE

insert_cassette("hello")

#> CrulAdapter enabled!

#> HttrAdapter enabled!

#> <vcr - Cassette> hello
#>   Record method: once
#>   Serialize with: yaml
#>   Persist with: FileSystem
#>   Re-record interval (s): 
#>   Clean outdated interactions?: FALSE
#>   update_content_length_header: FALSE
#>   decode_compressed_response: 
#>   allow_playback_repeats: FALSE
#>   allow_unused_http_interactions: 
#>   exclusive: 
#>   preserve_exact_body_bytes: FALSE

(x <- current_cassette())

#> <vcr - Cassette> stuffthings
#>   Record method: once
#>   Serialize with: yaml
#>   Persist with: FileSystem
#>   Re-record interval (s): 
#>   Clean outdated interactions?: FALSE
#>   update_content_length_header: FALSE
#>   decode_compressed_response: 
#>   allow_playback_repeats: FALSE
#>   allow_unused_http_interactions: 
#>   exclusive: 
#>   preserve_exact_body_bytes: FALSE

# by default does current cassette
x <- eject_cassette()

#> ejecting cassette: stuffthings

#> CrulAdapter disabled!

#> HttrAdapter disabled!

x

#> <vcr - Cassette> stuffthings
#>   Record method: once
#>   Serialize with: yaml
#>   Persist with: FileSystem
#>   Re-record interval (s): 
#>   Clean outdated interactions?: FALSE
#>   update_content_length_header: FALSE
#>   decode_compressed_response: 
#>   allow_playback_repeats: FALSE
#>   allow_unused_http_interactions: 
#>   exclusive: 
#>   preserve_exact_body_bytes: FALSE

# can also select by cassette name
# eject_cassette(cassette = "hello")