request and response summary methods

request and response summary methods

request_summary(request, request_matchers = "")

response_summary(response)

Arguments

request
a Request object

request_matchers
(character) a vector of matchers. Default: ""

response
a VcrResponse object

Value

character string, of either request or response

Details

By default, method and uri are included in the request summary - if body and/or headers are specified in request_matchers, then they are also included

HTTP status code and response body are included in the response summary. The response body is truncated to a max of 80 characters

Examples

# request
url <- "https://httpbin.org"
body <- list(foo = "bar")
headers <- list(
  `User-Agent` = "r-curl/3.2",
  `Accept-Encoding` = "gzip, deflate",
  Accept = "application/json"
)

(x <- Request$new("POST", url, body, headers))

#> <Request>
#>   Public:
#>     body: foo=bar
#>     clone: function (deep = FALSE) 
#>     from_hash: function (hash) 
#>     hash: NULL
#>     headers: list
#>     host: httpbin.org
#>     initialize: function (method, uri, body, headers, opts) 
#>     method: post
#>     opts: NULL
#>     path: NA
#>     query: NA
#>     scheme: https
#>     skip_port_stripping: FALSE
#>     to_hash: function () 
#>     uri: https://httpbin.org/
#>   Private:
#>     parsed_uri: function (uri) 
#>     without_standard_port: function (uri) 

request_summary(request = x)

#> [1] "post https://httpbin.org/"

request_summary(request = x, c('method', 'uri'))

#> [1] "post https://httpbin.org/"

request_summary(request = x, c('method', 'uri', 'body'))

#> [1] "post https://httpbin.org/ foo=bar"

request_summary(request = x, c('method', 'uri', 'headers'))

#> [1] "post https://httpbin.org/ r-curl/3.2 gzip, deflate application/json"

request_summary(request = x, c('method', 'uri', 'body', 'headers'))

#> [1] "post https://httpbin.org/ foo=bar r-curl/3.2 gzip, deflate application/json"

# response
status <- list(status_code = 200, message = "OK",
  explanation = "Request fulfilled, document follows")
headers <- list(
  status = "HTTP/1.1 200 OK",
  connection = "keep-alive",
  date = "Tue, 24 Apr 2018 04:46:56 GMT"
)
response_body <- "{\"args\": {\"q\": \"stuff\"}, \"headers\": {\"Accept\": \"application/json\"}}\n"
(x <- VcrResponse$new(status, headers,
   response_body, "HTTP/1.1 200 OK"))

#> <VcrResponse>
#>   Public:
#>     adapter_metadata: NULL
#>     body: {"args": {"q": "stuff"}, "headers": {"Accept": "applicat ...
#>     clone: function (deep = FALSE) 
#>     content_encoding: function () 
#>     delete_header: function (key) 
#>     edit_header: function (key, value = NULL) 
#>     from_hash: function (hash) 
#>     get_header: function (key) 
#>     hash: NULL
#>     headers: list
#>     http_version: 1.1
#>     initialize: function (status, headers, body, http_version, opts, adapter_metadata = NULL) 
#>     is_compressed: function () 
#>     opts: NULL
#>     status: list
#>     to_hash: function () 
#>     update_content_length_header: function () 

response_summary(x)

#> [1] "200 {\"args\": {\"q\": \"stuff\"}, \"headers\": {\"Accept\": \"application/json\"}} "